adding react router and protecting todo routes, adding log out

+++ server/controllers/TodoController.js
@@ -2,7 +2,10 @@ const Todo = require('../models/todo');
 const mongoose = require('mongoose');
 
 exports.getAllTodos = async (req, res) => {
-    userId = req.params.id;
+    if (!req.session.userId) {
+        return res.status(401).json({ message: 'Unauthorized' });
+    }
+    userId = req.session.userId;
     try {
         const count = await Todo.countDocuments({user : userId});
         if (count === 0) {
@@ -19,12 +22,16 @@ exports.getAllTodos = async (req, res) => {
 }
 
 exports.postCreateTodo = async (req, res) => {
+    if (!req.session.userId) {
+        return res.status(401).json({ message: 'Unauthorized' });
+    }
     try {
-        const { title, description, _id } = req.body;
+        const { title, description} = req.body;
+        const userId = req.session.userId;
         const todo = new Todo({
             title: title,
             description: description,
-            user: _id,
+            user: userId,
         });
         await todo.save();
         res.status(200).json(todo);
@@ -35,6 +42,9 @@ exports.postCreateTodo = async (req, res) => {
 
 
 exports.deleteTodo = async (req, res) => {
+    if (!req.session.userId) {
+        return res.status(401).json({ message: 'Unauthorized' });
+    }
     try {
         const todoId  = req.params.id;
         const todo = await Todo.findByIdAndDelete(todoId);
@@ -45,6 +55,9 @@ exports.deleteTodo = async (req, res) => {
 }
 
 exports.updateTodo = async (req, res) => {
+    if (!req.session.userId) {
+        return res.status(401).json({ message: 'Unauthorized' });
+    }
     try {
         const { id } = req.params;
         const { title, description, completed } = req.body;
